unicomer/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
google-cloud-documentai==2.3.0